Fix linter

+++ internal/provider/projects.go
@@ -10,7 +10,7 @@ import (
 )
 
 // Projects ...
-func (p *Provider) Projects(ctx context.Context, q string, limit, offset int) ([]*models.Project, uint64, error) {
+func (p *Provider) Projects(ctx context.Context, q string, limit, offset uint64) ([]*models.Project, uint64, error) {
 	projects, total, err := p.storage.Projects(ctx, q, limit, offset)
 	if err != nil {
 		return nil, 0, fmt.Errorf("storage.Projects: %w", err)
